Update db_Get_field_list.4dm
Bug in method db_Get_field_list, in line 143 and 144 it’s always appending the fields regardless of what table. Then in line 154 it loops for all the fields but only checking a specific table. It will have an issue if the table have relationships with a lot of tables.